完成热部署devtools

+++ could2020/cloud-provider-payment8001/src/main/java/com/it/springcloud/dao/PaymentDao.java
@@ -8,5 +8,5 @@ public interface PaymentDao {
 
     public int add(Payment payment);
 
-    public Payment getPaymentById(Integer id);
+    public Payment getPaymentById(Long id);
 }
